chore: Update dependencies and add emoji support
This commit updates the package version to 1.0.5, upgrades React and related packages to version 19.1.1, and adds new dependencies including react-live and emoji support. Additionally, it introduces a new emoji.json file for emoji representation and modifies the i18n configuration for improved language detection. The App component is refactored to enhance the code structure and default preview visibility.